fix: update balance on claims without reload
